ProductionOrder & Product classes Refactored Resource - it's float now (it's much easier that way) Added 2 new stats (Production time & Cost)

+++ Assets/Scripts/MapElements/Resource.cs
@@ -6,10 +6,10 @@ namespace MechWars.MapElements
     {
         public override bool Interactible { get { return true; } }
 
-        public int value;
-        int startValue;
+        public float value;
+        float startValue;
 
-        public float Size {  get { return (float)value / startValue; } }
+        public float Size {  get { return value / startValue; } }
 
         public Resource()
         {
